news_producer/requirements.txt
kafka-python
feedparser